Fix Enum Serialization
Named enums were being incorrectly serialized via the derive macro. This was causing the value of the enum to be disregarded. This change implements `Serialize` and `Deserialize` explicitly for enums.

Additionally, a new protocol test was added that avoid using sets–this allow it to be run without being flaky until #37 is fixed.