Add restJson error parsing & prefix headers support (#274)
* restJson1 Header Deserialization Support

This adds support for the `@httpHeader` trait during request deserialization. This is facilitated via a `Read` abstraction which allows parsers to consume part of an input and return the remaining input. We use this to enable things like `HttpDates` which actually include commas to be parsed in a comma delimited fashion.

The internal libraries may eventually be refactored to reduce the amount of generics if it proves to be a compiler bottleneck.

* Add support for parsing request bodies

* Add response code parsing

* Lots of refactorings to restJson response parsing

* Fix some clippy lints

* Add document deserialization support and delete the failing tests

* Add new error handling (all RestJson protocol tests passing!)

* Add prefix headers parsing support

* Fix Rust Runtime clippy warnings

* Cleanup some more dead code / clippy fixes

* Update external functions to take HeaderMap instead of the value iterator

* Hard error on unreconized binding location

* Refactor parse function generation to be on-demand

* One more round of refactoring

* Few more cleanups