Send Poll::Ready(None) on empty body (#325)
When used over H/2, sending an empty data frame for a GET can trigger some problematic behavior in Hyper. Hyper should probably handle this, but this ensures that we will
not hit this behavior.

When sending data to API Gateway services, we are getting GOAWAY because this causes to send duplicate empty data frames.