Parser Generator for XmlBindingTraits
This diff adds two components:
- `smithy-xml`: A scoped-reader XML abstraction on top of `xmlparser`
- `XmlBindingTraitParserGenerator`: A code generator for deserializing data targetted with the XmlBinding traits

This is not currently used outside of the tests, but this implementation passes all protocol tests expect for escaping (wiring is still WIP).

This commit also adds two inlinable abstractions for error parsing which can be swapped in at code generation time based on the service traits.