Gracefully handle alternate error message fields (#374)
Errors returned by AWS services may be stored at `Message`, `message` or `errorMessage`. This updates the relevant code to search all three locations when surfacing a clean error message to the customer. This diff also fixes a long-standing shortcoming of SdkError where it failed to provide a clean Display implementation.