add initial profile file provider support (#640)
* add initial profile file provider support

This commit establishes baseline support for profile file providers. The only named credential source currently supported is `Environment` but we'll be adding more sources in the future. To support testing, the `dvr` module was added to smithy-client. There are some limitations in this which will be addressed going forward, primarily around time control so we can do request verification.

* Fix missed rename

* CR feedback: add doc comments, cleanup tests

* Improve take requests API

* Fix missed refactoring

* Add missing test data

* Fix intradoc link

* Update changelog

* resolve clippy errors