Add support for web identity token providers (#654)
* Add support for web identity token providers

This diff adds support for WebIdentity token providers which may be utilized in three ways:
1. Via environment variables
2. Via a profile directly
3. As a source profile of another assume role profile

Along the way, I overhauled the way that tests work to be primarily self-contained test directories.

* make fs::read_to_end async

For future compatibility I updated read_to_end to be an async function. This drove a subsequent change to push profile file parsing into credentials loading. This should be generally useful anyway since we'll want a way to force reparsing of the profile file.

* Revert change to env

This removes the need for an allocation in the non-test case.

* Fix broken doc tests

* Fix clippy

* Update changelog, fix docs

* Remove caching from profile provider