Fix signing when Hyper ALPN negotiates to h2
The endpoint middleware was adding a host header to the request, and
this worked when using HTTP/1.1. However, when the connection gets
upgraded to h2, this results in the remote service expecting the
canonical request to have two `host` headers.

This change moves the addition of the host header into the signing code,
and lets hyper take charge of the host header.