Fix errors discovered with RestJson protocol tests (#926)
This PR fixes some errors discovered running the protocol tests for RestJson. It allows to fully build the RestJson protocol tests model, but it doesn't still generate proper tests (IE cargo test will fail) since we still haven't implemented parts for the protocol, like httpQuery trait.

The main problem discovered and fixed is that we were generating JSON deserializers for input structures without a JSON body (for traits like httpQuery that we still do not support). This was causing the in memory deduplication to get confused and the only visible effect was some clippy warning around mutable builders that shouldn't have been mutable. Without the protocol tests this would not surface since the ebs and simple model do not exercise that specific code path in the codegen.

Signed-off-by: Bigo <1781140+crisidev@users.noreply.github.com>