`aws-smithy-http-server`: remove S3-like URI pattern conflict avoidance test (#1070)
The previous commit introduced a failing test because CI doesn't have
Rust tests as a blocking step and auto-merge was enabled.

With the two operations:

1. `ListBuckets` with URI pattern `/`; and
2. `ListObjects` with URI pattern `/{bucket}`,

the test wants to route the request `/` to the first operation. However,
in #1029 we started allowing empty path segments. Since `ListBuckets`
has more weight than `ListObjects`, the request matches against it,
binding `""` to the `bucket` label.