Fix SDK assemble Gradle pathing issue (#1343)
On M1 Macs, Gradle's working directory is the path of the Gradle daemon
rather than the project root. The SDK's `IntegrationTestDecorator` was
assuming that the working directory would always be the project root, so
assemble was failing on this setup.

This commit adds SDK-specific settings so that the integration test
directory can be configured in `smithy-build.json`.