Remove TODO about required `httpHeader`-bound collection shapes
We do support required `httpHeader`-bound collection shapes, as
evidenced by the test operation this commit adds, which is generated by
the server, that does not use `Option` for `required` member shapes.

This is because even though the deserialization always returns an
optional collection, we return `None` if no headers are found (as
opposed to an empty collection), so the call that sets the member field
on the operation structure does nothing; when `build()` gets called, a
`BuildError` is raised because the member is required.