Fix release issues by introducing backwards compat into `sdk-versioner` (#1434)
The `sdk-sync` tool is running into issues during release due to the
`sdk-versioner` refactor commit since it is using a Docker build image
with the tools precompiled into it. Either the `sdk-versioner` binary is
too new and can't code generate the old commits, or it's too old and
can't generate the new commits.

This reintroduces the old `sdk-versioner` behavior so that it can
succeed against both sets of commits to get a release out.