Try to move MaybeConstrained, Constrained into aws-smithy-http-server
But it's not possible, because if we do, then we need both
`RuntimeType`s to declare a dependency on
`ServerCargoDependency.SmithyHttpServer`, but we can't do that in
`RuntimeType.kt`, which is in `codegen-core`, since we can't make
`codegen-core` take a dependency on `codegen-server`.

But if we put them in `ServerRuntimeType`, we need to move the `data
class MaybeConstrained` (which would move the `MaybeConstrained`
`RuntimeType`) out of `RustTypes`, since `RustTypes` is in in
`codegen-core`. But `RustTypes` is `sealed`, and we can't inherit from a
sealed class outside the module where the class is defined. We'd have to
make `RustTypes` `abstract`. But then `RustType.render` can't be
exhaustive, and I don't know how to continue.